agera_one v0.0.1 AgeraOneWeb.Router.Helpers

Module with named helpers generated from AgeraOneWeb.Router.

Functions

abi_path(conn_or_endpoint, action)

abi_path(conn_or_endpoint, action, params)

abi_path(conn_or_endpoint, action, id, params)

abi_url(conn_or_endpoint, action)

abi_url(conn_or_endpoint, action, params)

abi_url(conn_or_endpoint, action, id, params)

balance_path(conn_or_endpoint, action)

balance_path(conn_or_endpoint, action, params)

balance_path(conn_or_endpoint, action, id, params)

balance_url(conn_or_endpoint, action)

balance_url(conn_or_endpoint, action, params)

balance_url(conn_or_endpoint, action, id, params)

block_path(conn_or_endpoint, action)

block_path(conn_or_endpoint, action, params)

block_path(conn_or_endpoint, action, id, params)

block_url(conn_or_endpoint, action)

block_url(conn_or_endpoint, action, params)

block_url(conn_or_endpoint, action, id, params)

path(data, path)

Generates the path information including any necessary prefix.

rpc_path(conn_or_endpoint, action)

rpc_path(conn_or_endpoint, action, params)

rpc_url(conn_or_endpoint, action)

rpc_url(conn_or_endpoint, action, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

statistics_path(conn_or_endpoint, action)

statistics_path(conn_or_endpoint, action, params)

statistics_url(conn_or_endpoint, action)

statistics_url(conn_or_endpoint, action, params)

transaction_path(conn_or_endpoint, action)

transaction_path(conn_or_endpoint, action, params)

transaction_path(conn_or_endpoint, action, id, params)

transaction_url(conn_or_endpoint, action)

transaction_url(conn_or_endpoint, action, params)

transaction_url(conn_or_endpoint, action, id, params)

url(data)

Generates the connection/endpoint base URL without any path information.